encryption is working?

+++ Encryption.java
@@ -1,5 +1,5 @@
 public class Encryption {
 
-public native int[] encrypt(byte[] value, int[] key);
+public native int[] encrypt(int[] value, int[] key);
 
 }
